tsp-main/podmanager/requirements.txt
gunicorn
Flask
pyyaml
pysqlite3
requests
kubernetes
Jinja2==2.9.4
MarkupSafe==0.23
Werkzeug==0.11.15
wheel==0.24.0
itsdangerous==2.0.1
